Check if boot path could be found

+++ tests/lib/state.sh
@@ -7,6 +7,8 @@ SNAPD_STATE_PATH="$SPREAD_PATH/snapd-state"
 
 save_all_snap_state() {
     local boot_path="$(get_boot_path)"
+    test -n "$boot_path" || return 1
+
     mkdir -p "$SNAPD_STATE_PATH"
     cp -rf /var/lib/snapd "$SNAPD_STATE_PATH/snapdlib"
     cp -rf "$boot_path" "$SNAPD_STATE_PATH/boot"
@@ -17,6 +19,8 @@ restore_all_snap_state() {
     # we need to ensure that we also restore the boot environment
     # fully for tests that break it
     local boot_path="$(get_boot_path)"
+    test -n "$boot_path" || return 1
+
     rm -rf /var/lib/snapd/*
     cp -rf "$SNAPD_STATE_PATH"/snapdlib/* /var/lib/snapd
     cp -rf "$SNAPD_STATE_PATH"/boot/* "$boot_path"
